+++ software/distro/setup/base-os/setup.sh
@@ -42,6 +42,14 @@ else
   panic "$description"
 fi
 
+description="download Forklift & Forklift pallet"
+report_starting "$description"
+if $build_scripts_root/forklift/install.sh ; then
+  report_finished "$description"
+else
+  panic "$description"
+fi
+
 description="configure Raspberry Pi-specific hardware"
 report_starting "$description"
 if $build_scripts_root/platform-hardware/config.sh ; then
@@ -62,14 +70,6 @@ fi
 description="configure networking"
 report_starting "$description"
 if $build_scripts_root/networking/config.sh ; then
-  report_finished "$description"
-else
-  panic "$description"
-fi
-
-description="remove unnecessary artifacts"
-report_starting "$description"
-if $build_scripts_root/cleanup/clean.sh ; then
   report_finished "$description"
 else
   panic "$description"