+++ software/distro/setup/planktoscope-app-env/gpio-stepper-init/config.sh
@@ -9,8 +9,3 @@ config_files_root=$(dirname $(realpath $BASH_SOURCE))
 file="/etc/systemd/system/planktoscope-org.init-gpio-steppers.service"
 sudo cp "$config_files_root$file" "$file"
 sudo systemctl enable planktoscope-org.init-gpio-steppers.service
-
-# Add shell script for initializing stepper motors
-mkdir -p $HOME/.local/bin
-file="$HOME/.local/bin/release-gpio-steppers.sh"
-cp "$config_files_root$file" "$file"