+++ software/distro/setup/planktoscope-app-env/networking/config.sh
@@ -14,52 +14,32 @@ sudo bash -c "echo \"$new_hostname\" > /etc/hostname"
 sudo sed -i "s/^127\.0\.1\.1.*$current_hostname$/127.0.1.1\t$new_hostname/g" /etc/hosts
 sudo hostnamectl set-hostname "$new_hostname"
 
-# Set the default SSID for the self-hosted wifi network, which will be updated with the machine name on boot
-file="/etc/hostapd/hostapd-ssid-autogen-warning.snippet"
-# This sed command uses `~` instead of `/` because the warning comments also include `/` characters.
-# The awk subcommand is needed to escape newlines for sed.
-sudo sed -i "s~^ssid=.*$~$(awk '{printf "%s\\n", $0}' $file)ssid=pkscope~g" /etc/hostapd/hostapd.conf
-
 # Download tool to generate machine names based on serial numbers
 machinename_version="0.1.3"
 curl -L "https://github.com/PlanktoScope/machine-name/releases/download/v$machinename_version/machine-name_${machinename_version}_linux_arm.tar.gz" \
-  | tar -xz -C $HOME/.local/bin/ machine-name
+  | sudo tar -xz -C /usr/bin/ machine-name
+sudo mv /usr/bin/machine-name "/usr/bin/machine-name-${machinename_version}"
+sudo ln -s "/usr/bin/machine-name-${machinename_version}" /usr/bin/machine-name
 
 # Automatically generate the machine name and write it to a file upon boot
-mkdir -p $HOME/.local/etc
-file="$HOME/.local/etc/machine-name"
-cp "$config_files_root$file" "$file"
-mkdir -p $HOME/.local/bin
-file="$HOME/.local/bin/update-machine-name.sh"
-cp "$config_files_root$file" "$file"
 file="/etc/systemd/system/planktoscope-org.update-machine-name.service"
 sudo cp "$config_files_root$file" "$file"
 sudo systemctl enable planktoscope-org.update-machine-name.service
 
 # Automatically update the SSID upon creation of the self-hosted wifi network based on the machine name
-mkdir -p $HOME/.local/etc/hostapd
-file="$HOME/.local/etc/hostapd/ssid.snippet"
-cp "$config_files_root$file" "$file"
-mkdir -p $HOME/.local/bin
-file="$HOME/.local/bin/update-ssid-machine-name.sh"
-cp "$config_files_root$file" "$file"
-file="/etc/systemd/system/planktoscope-org.update-ssid-machine-name.service"
+file="$config_files_root/etc/hostapd/hostapd-ssid-autogen-warning.snippet"
+# This sed command uses `~` instead of `/` because the warning comments also include `/` characters.
+# The awk subcommand is needed to escape newlines for sed.
+sudo sed -i "s~^ssid=\(.*\)$~$(awk '{printf "%s\\n", $0}' $file)ssid=\\1~g" /etc/hostapd/hostapd.conf
+file="/etc/systemd/system/planktoscope-org.update-hostapd-ssid-machine-name.service"
 sudo cp "$config_files_root$file" "$file"
-sudo systemctl enable planktoscope-org.update-ssid-machine-name.service
+sudo systemctl enable planktoscope-org.update-hostapd-ssid-machine-name.service
 
 # Change Cockpit settings
 file="/etc/cockpit/cockpit.conf"
 sudo bash -c "cat \"$config_files_root$file.snippet\" >> \"$file\""
 
 # Automatically update the Cockpit origins upon boot with the machine name
-mkdir -p $HOME/.local/etc/cockpit
-file="$HOME/.local/etc/cockpit/origins-base.snippet"
-cp "$config_files_root$file" "$file"
-file="$HOME/.local/etc/cockpit/origins-machine-name.snippet"
-cp "$config_files_root$file" "$file"
-mkdir -p $HOME/.local/bin
-file="$HOME/.local/bin/update-cockpit-origins-machine-name.sh"
-cp "$config_files_root$file" "$file"
 file="/etc/systemd/system/planktoscope-org.update-cockpit-origins-machine-name.service"
 sudo cp "$config_files_root$file" "$file"
 sudo systemctl enable planktoscope-org.update-cockpit-origins-machine-name.service
@@ -67,28 +47,11 @@ sudo systemctl enable planktoscope-org.update-cockpit-origins-machine-name.servi
 # Change dnsmasq settings
 file="/etc/dnsmasq.d/planktoscope.conf"
 sudo cp "$config_files_root$file" "$file"
-mkdir -p $HOME/.local/etc
-file="$HOME/.local/etc/hosts-autogen-warning.snippet"
-cp "$config_files_root$file" "$file"
-file="$HOME/.local/etc/hosts-base.snippet"
-cp "$config_files_root$file" "$file"
-cp "$HOME/.local/etc/hosts-autogen-warning.snippet" \
-  "$HOME/.local/etc/hosts"
-cat "$HOME/.local/etc/hosts-base.snippet" \
-  >> "$HOME/.local/etc/hosts"
-
-# Automatically update hostnames upon boot with the machine name
-mkdir -p $HOME/.local/etc
-file="$HOME/.local/etc/hosts-machine-name.snippet"
-cp "$config_files_root$file" "$file"
-mkdir -p $HOME/.local/bin
-file="$HOME/.local/bin/update-hosts-machine-name.sh"
-cp "$config_files_root$file" "$file"
 file="/etc/systemd/system/planktoscope-org.update-hosts-machine-name.service"
 sudo cp "$config_files_root$file" "$file"
 sudo systemctl enable planktoscope-org.update-hosts-machine-name.service
-file="$HOME/.local/bin/update-hostname-machine-name.sh"
-cp "$config_files_root$file" "$file"
+
+# Automatically update system hostname upon boot with the machine name
 file="/etc/systemd/system/planktoscope-org.update-hostname-machine-name.service"
 sudo cp "$config_files_root$file" "$file"
 sudo systemctl enable planktoscope-org.update-hostname-machine-name.service