+++ software/distro/setup/planktoscope-app-env/node-red-frontend/install.sh
@@ -22,11 +22,6 @@ pip3 install smbus2==0.4.3
 curl -sL https://raw.githubusercontent.com/node-red/linux-installers/master/deb/update-nodejs-and-nodered \
   | bash -s - --confirm-install --confirm-pi --no-init
 
-# Create a settings file to run Node-RED
-file="$HOME/.node-red/settings.js"
-cp "$config_files_root$file" "$file"
-sudo chown 0:0 "$file"
-
 # Add systemd service modification to make Node-RED wait until Mosquitto has started
 # FIXME: The Node-RED frontend should instead be fixed so that it does not need to wait until
 # Mosquitto has started in order to work - until Mosquitto is up, the frontend should display a