+++ software/distro/setup/planktoscope-app-env/setup.sh
@@ -37,9 +37,9 @@ function panic {
 
 # Run sub-scripts
 
-description="enable CPU overclocking"
+description="configure networking"
 report_starting "$description"
-if $build_scripts_root/overclocking/install.sh ; then
+if $build_scripts_root/networking/config.sh ; then
   report_finished "$description"
 else
   panic "$description"
@@ -85,25 +85,9 @@ else
   panic "$description"
 fi
 
-description="configure networking"
+description="enable CPU overclocking"
 report_starting "$description"
-if $build_scripts_root/networking/config.sh ; then
-  report_finished "$description"
-else
-  panic "$description"
-fi
-
-description="download Forklift pallet"
-report_starting "$description"
-if $build_scripts_root/forklift/install.sh ; then
-  report_finished "$description"
-else
-  panic "$description"
-fi
-
-description="remove unnecessary artifacts"
-report_starting "$description"
-if $build_scripts_root/cleanup/clean.sh ; then
+if $build_scripts_root/overclocking/install.sh ; then
   report_finished "$description"
 else
   panic "$description"